requirements.txt
# Telegram Bot
aiogram==3.3.0

# API Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# Speech Recognition
openai-whisper==20231117

# Date Parsing
dateparser==1.2.0

# Scheduler
apscheduler==3.10.4

# Environment
python-dotenv==1.0.0
